replace visible string tag `*` with `_`
Due to this breaking change you have to replace all occurrences of
`*` with `_`.

The motivation for this change is that this program can be also used
for little calculation excises, e.g. `- 5 _* 3 = 15_` which appears
as `_ * 3 = 15` on the screen. In these expressions `*` can not be
a special character so we use `_` instead.